+++ flexmeasures/api/v1_2/implementations.py
@@ -28,7 +28,6 @@ from flexmeasures.api.common.utils.validators import (
     type_accepted,
     assets_required,
     optional_duration_accepted,
-    usef_roles_accepted,
     units_accepted,
     parse_isodate_str,
 )
@@ -125,7 +124,6 @@ def get_device_message_response(generic_asset_name_groups, duration):
     return dict(**response, **d), s
 
 
-@usef_roles_accepted("Prosumer")
 @type_accepted("PostUdiEventRequest")
 @units_accepted("State of charge", "kWh", "MWh")
 @as_json