+++ flexmeasures/data/services/forecasting.py
@@ -30,7 +30,7 @@ from flexmeasures.utils.time_utils import (
 """
 The life cycle of a forecasting job:
 1. A forecasting job is born in create_forecasting_jobs.
-2. It is run in make_forecasts which writes results to the db.
+2. It is run in make_rolling_viewpoint_forecasts or make_fixed_viewpoint_forecasts, which write results to the db.
    This is also where model specs are configured and a possible fallback model is stored for step 3.
 3. If an error occurs (and the worker is configured accordingly), handle_forecasting_exception comes in.
    This might re-enqueue the job or try a different model (which creates a new job).
@@ -98,7 +98,7 @@ def create_forecasting_jobs(
     jobs: List[Job] = []
     for horizon in horizons:
         job = Job.create(
-            make_forecasts,
+            make_rolling_viewpoint_forecasts,
             kwargs=dict(
                 asset_id=asset_id,
                 timed_value_type=timed_value_type,
@@ -122,7 +122,7 @@ def create_forecasting_jobs(
     return jobs
 
 
-def make_forecasts(
+def make_fixed_viewpoint_forecasts(
     asset_id: int,
     timed_value_type: str,
     horizon: timedelta,
@@ -130,10 +130,30 @@ def make_forecasts(
     end: datetime,
     custom_model_params: dict = None,
 ) -> int:
+    """Build forecasting model specs, make fixed-viewpoint forecasts, and save the forecasts made.
+
+    Each individual forecast is a belief about a time interval.
+    Fixed-viewpoint forecasts share the same belief time.
+    See the timely-beliefs lib for relevant terminology.
     """
-    Build forecasting model specs, make rolling forecasts, save the forecasts made.
-    Each individual forecast is a belief about an interval.
-    Returns the number of forecasts made.
+    # todo: implement fixed-viewpoint forecasts
+    raise NotImplementedError
+
+
+def make_rolling_viewpoint_forecasts(
+    asset_id: int,
+    timed_value_type: str,
+    horizon: timedelta,
+    start: datetime,
+    end: datetime,
+    custom_model_params: dict = None,
+) -> int:
+    """Build forecasting model specs, make rolling-viewpoint forecasts, and save the forecasts made.
+
+    Each individual forecast is a belief about a time interval.
+    Rolling-viewpoint forecasts share the same belief horizon (the duration between belief time and knowledge time).
+    Model specs are also retrained in a rolling fashion, but with its own frequency set in custom_model_params.
+    See the timely-beliefs lib for relevant terminology.
 
     Parameters
     ----------
@@ -150,6 +170,8 @@ def make_forecasts(
     :param custom_model_params: dict
         pass in params which will be passed to the model specs configurator,
         e.g. outcome_var_transformation, only advisable to be used for testing.
+    :returns: int
+        the number of forecasts made
     """
     # https://docs.sqlalchemy.org/en/13/faq/connections.html#how-do-i-use-engines-connections-sessions-with-python-multiprocessing-or-os-fork
     db.engine.dispose()
@@ -267,7 +289,7 @@ def handle_forecasting_exception(job, exc_type, exc_value, traceback):
     if "fallback_model_search_term" in job.meta:
         if job.meta["fallback_model_search_term"] is not None:
             new_job = Job.create(
-                make_forecasts,
+                make_rolling_viewpoint_forecasts,
                 args=job.args,
                 kwargs=job.kwargs,
                 connection=current_app.queues["forecasting"].connection,